move get by id tests to queries
