+++ backend/src/domain/repositories/recipe/postgres.rs
@@ -332,6 +332,3 @@ impl PostgresRecipeRepository {
         Arc::new(Box::new(self))
     }
 }
-
-#[cfg(test)]
-mod tests;
